Refactor: Move calendar sensor class to submodule
In prepration to have more than one type of sensor the current
RentalControlCalSensor needed to move to a submodule for easier
development.
