+++ custom_components/rental_control/sensors/__init__.py
@@ -1,0 +1,1 @@
+"""Sensors submodule."""
